Refactored "util" folder to "core"
