+++ src/client/java/com/fexl/circumnavigate/CircumnavigateClient.java
@@ -4,7 +4,7 @@ package com.fexl.circumnavigate;
 
 import com.fexl.circumnavigate.client.storage.TransformersStorage;
 import com.fexl.circumnavigate.network.packet.ClientboundWrappingDataPacket;
-import com.fexl.circumnavigate.util.WorldTransformer;
+import com.fexl.circumnavigate.core.WorldTransformer;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.core.registries.Registries;
