+++ src/client/java/com/fexl/circumnavigate/mixin/client/chunkHandle/OptionsMixin.java
@@ -2,7 +2,7 @@
 
 package com.fexl.circumnavigate.mixin.client.chunkHandle;
 
-import com.fexl.circumnavigate.util.WorldTransformer;
+import com.fexl.circumnavigate.core.WorldTransformer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.OptionInstance;
 import net.minecraft.client.Options;
@@ -21,7 +21,7 @@ public abstract class OptionsMixin {
 	@Inject(method = "getEffectiveRenderDistance", at = @At("HEAD"), cancellable = true)
 	public void getEffectiveRenderDistance(CallbackInfoReturnable<Integer> cir) {
 		WorldTransformer transformer = Minecraft.getInstance().level.getTransformer();
-		//System.out.println("Server: " + accessor.getServerRenderDistance() + " Client: " + accessor.getRenderDistance().get());
+
 		int renderDistance = accessor.getServerRenderDistance() > 0 ? Math.min(accessor.getRenderDistance().get(), accessor.getServerRenderDistance()) : accessor.getRenderDistance().get();
 		cir.setReturnValue(transformer.limitViewDistance(renderDistance));
 	}
