+++ src/client/java/com/fexl/circumnavigate/mixin/client/debug/ChunkBorderRendererMixin.java
@@ -2,10 +2,9 @@
 
 package com.fexl.circumnavigate.mixin.client.debug;
 
-import com.fexl.circumnavigate.util.WorldTransformer;
+import com.fexl.circumnavigate.core.WorldTransformer;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.debug.ChunkBorderRenderer;
