+++ src/client/java/com/fexl/circumnavigate/mixin/client/packet/ServerboundMovePlayerPacketMixin.java
@@ -2,7 +2,7 @@
 
 package com.fexl.circumnavigate.mixin.client.packet;
 
-import com.fexl.circumnavigate.util.WorldTransformer;
+import com.fexl.circumnavigate.core.WorldTransformer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.protocol.game.ServerboundMovePlayerPacket;
 import org.spongepowered.asm.mixin.Mixin;
