+++ src/main/java/com/fexl/circumnavigate/core/Distance.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: AGPL-3.0-only */
 
-package com.fexl.circumnavigate.util;
+package com.fexl.circumnavigate.core;
 
 import net.minecraft.world.level.chunk.LevelChunkSection;
 import net.minecraft.world.phys.Vec2;
